refactor(plants): move services to services/ directory (step 4)
- Move plant.service.ts from plant-service/ to services/
- Move image.service.ts from image-service/ to services/
- Update all imports across components and dialogs
- Remove empty plant-service/ and image-service/ directories
