+++ src/app/plants/services/image.service.spec.ts
