+++ src/app/plants/services/image.service.ts
