+++ src/app/plants/services/plant.service.ts
